add move

+++ client/src/main/java/com/shot_tank/models/Location.java
@@ -2,10 +2,10 @@ package com.shot_tank.models;
 
 public class Location {
 
-    public int x, y;
+    public double x, y;
     public double angle;
     
-    public Location(int x, int y, double angle) {
+    public Location(double x, double y, double angle) {
         this.x = x;
         this.y = y;
         this.angle = angle;
